cliente controller, pagina homeCliente ( html en rama 9 )

+++ src/main/java/Servicios/UsuarioService.java
@@ -45,7 +45,7 @@ import modelo.Usuario;
 				Session.rollbackTransaction();
 			}
 		}
-		public static Usuario obtenerUsuarioPorId(int id)
+		public static Usuario obtenerUsuarioPorId(Long id)
 		{
 		return Session.getSession().find(Usuario.class,id);
 		}
